src: Rename `bindings.rs` to `raw_bindings.rs`
