+++ src/lib.rs
@@ -1,5 +1,5 @@
-mod bindings;
-pub use bindings::*;
+mod raw_bindings;
+pub use raw_bindings::*;
 
 
 /**
